Added new cube array deletion and undo button
The game now uses RaycastAll() in order to see all the cubes that were
hit. Will use this later to be able to hold down the left mouse button
and keep deleting cubes. Also added a list of the deleted cubes that can
be brought back into existence by hitting the Z key. Mostly going to use
that for debugging and possibly puzzle creation.