Hit Cubes are sorted and rotate with RMB
All the cubes that are hit by the Raycast upon clicking the left mouse
button are now sorted from closest to furthest from the camera. Will use
this to determine which are in the same row/column as the closest, then
allow the player to keep deleting that row/column until none are left.
To keep cube deletiong and rotation distinct for now, the player holds
and uses the right mouse button to rotate the puzzle. This might be a
reversable option in the future.