Added hold LMB to delete rows
This rapid row deletion needs refining because at the moment it gets
diagonally connected cubes, not just the ones in the same row as the
initially clicked cube. Also need to make it delete cubes at a slower
pace so that the player can stop it early if they want to.