Sliders being draggable on one axis working
I still need to check if the sliders still work at weird angles, and if
the player drags too quickly they can move off the slider, so maybe make
it slide until the mouse button is released? Next up I have to link the
sliding movement to the hiding of the cubes.