Beta Proof of Concept v1
Separate types of cubes are now read in from the text file that contains
the schematic for the puzzle. Still need to have it set the numbers of
the cubes based on the layout of the puzzle, but at the moment there are
good and bad cubes whose layout can be read in. Issue with trying to
click on cubes through transparent ones, needs to be fixed.