Made the puzzle file checker a bit more robust
The puzzle text file checker now detects most issues that might occur
when reading in a file, such as too many rows in a layer or too many
cubes in a row. Need to double check that not too many layers can be
made, and other issues such as too few cubes in a row, too few layers,
etc.

Also made a 3 dimensional integer array that updates in parallel to the
format of the final cube array. Eventually these methods will coincide
or I'll just read this integer array to determine the contents of the
puzzle itself.