Made the puzzle file format checker even more robust
The puzzle file checker method now gives more detailed results on issues
with the puzzle file format, such as too few row in a layer, too many or
too few layers, not enough cubes in a row, etc. Can probably still be
improved but for now I'll leave it alone.