Testing out setting different faces of cubes
Made a work in progress texture that will be used to create the faces
for the cubes. At the moment just demoing a cube with different sides,
once that is up and running I'll try running through the entire array of
cubes and seeing how many are in each row and column to determine the
numbers for the faces of the cubes.