Cube face texture assignment is now working
The faces of the cube puzzle are now successfully assigned via a script
using a texture map. I need to go back into the texture map and realign
and re-center the numbers for the rows and columns. Also need to make
the dark texture darker to help differentiate it from the light texture.