Created methods to set a specific face to a texture
Created a method that will set an exact face to a given texture (1, 0,
dark 1, dark 0, etc.). This will be used with the cube initialization in
the CubeManager to set the starting state of the cubes. Will also create
an array that will store the states of each face of the cube to be
accessed when the cube gets highlighted or un-highlighted later.