Cubes now have their faces and numbers set properly
When the puzzle is created, each face of the cube now shows the proper
number of keyCubes in the same row as the face of the cube. The puzzle
is nearly concept-complete, just need to add a system for
highlighting/flagging cubes to not be deletable since the player thinks
they are part of the solution.