Cube faces are now set according to the 12x12 texture map
The 12x12 texture map now is being used to create the textures for each
face of each cube. Up next is making each texture face not only show the
number of keyCubes in a row/column, but also indicate if they are in
separate groups, via circles and squares around the numbers. Flagging
cubes is still in the pipeline for things to add.