New test puzzles added
Created a few text files that contain solutions to a few puzzles for
testing purposes. I need to go back in to the various methods that use
things like the puzzle size in specific dimensions and make sure they
work for asymmetrical puzzles. The current issue to tackle involving
this problem is the HideIndices array that was made just using a single
dimension. Might need to make one for each slider, or have the array be
re-initialized depending on the slider that is in use.