Hidden cubes' colliders are now disabled when hidden
The cubes that are "hidden" by sliders can now be clicked through by
disabling their colliders while hidden. Also adjusted the position of
the cubes in comparison to the main puzzle so that they don't overlap
with the cubes. Next up is highlighting the cubes sot hey can't be
deleted.