Cube flagging system is now in place
Feel free to click on flagged cubes, they cannot be deleted. If you
press left shift, the game changes to flagging mode and now when you
click on cubes they change color and cannot be deleted. This will help
players mark the cubes they think are part of the solution so they don't
have to worry about deleting them accidentally.