File reading to create puzzle selection UI started
A file reader has been created that can take in all the .txt files in
the StreamingAssets folder (the new home for puzzle solutions). These
will be read in and used to populate a list of puzzles that the player
can choose from. One hiccup that will need a bit of work in the future
is stuff relating to the android and web versions of the game.