Puzzles loaded when clicking on the buttons
The button that indicates a specific solution now loads its respective
puzzle when clicked on. The previous puzzle cubes and deleted cubes are
wiped clean to clear up the scene. Might need to use object pooling or
something instead of destroying, but normally you can't destroy and
create puzzle on the fly. Need to add a confirmation upon selecting a
new puzzle. Also need to fix the rotation/position of the sliders upon
slider creation.