Camera finally working as desired. Other bug fixes related to creation of new puzzles.
Camera finally working as intended. The puzzle rotates towards and away from the camera, on an axis that always points to the right of the camera's forward direction. This vertical rotation is clamped to +90 or -90 degrees away from the starting 0 rotation. Horizontal rotation is based on the vertical axis of the puzzle itself and is not clamped.
Also, puzzle selection now hides after a puzzle is loaded. Sliders don't rotate now when loading a new puzzle. A new puzzle resets the puzzle rotation back to the default (0,0,0).