Z-Layer new hiding method now works
The z-layer hiding is now based on units moved towards or away from its starting position, so that every time one whole unit of distance is reached, a layer is correspondingly hidden or shown. Next up is implementing that for the x-slider. I also need the dragging of the sliders to work when dragging away from the puzzle instead of just towards the puzzle.