Sliding is finally working as intended (for the most part)
By manipualting the localPosition of the slider in relation to the parent object, the slider can now only move along one axis. The only thing left is to let the player slide back from the closest and furthest possible positions. Maybe I can add a notch-like feeling to the movement so it corresponds to each layer hidden/revealed. This took so long, almost as long as the camera controls. Finally glad it worked.