The Sliders are now working
The sliders are now clamping their movement within the desired range. Thank you transform.InverseTransformPoint. Last thing I want to do with them is make the hidden slider reappear when the currently being used slider is manually returned to its starting position as opposed to double-clicking it to reset the sliders/layers.