Cube face highlighting in build mode is now working.
Thanks to some old code that I had left behind from setting the textures of cubes, I was able to add the functionality of changing just the color of the face of the cube that is under the mouse while in build mode. This will help the player know where the cube they are going to create will go. Next up is adding cube deletion to build mode, saving the cubes in a list that can be exported as a puzzle solution, and alternating cube colors in build mode to improve visibility.