Slider movement is working as intended.
The slider no longer drags off into infinity when used at certain parts of its path. One issue currently troubling the sliders is that if the user drags the slider too quickly, they can lose "progress" on the slider's movement and not have the proper layers be shown/hidden. I'll try going back to the old method of hiding layers based on the current position of the slider instead of its change in movement.