JSON files are able to be created and named now.
There is now a prompt that comes up when the player presses the S key to save a dummy JSON file. This will be called later when hitting the right key/UI element. Need to get the conversion of cubes made in Build Mode to save properly. Need to also test reading in and creating puzzles from the JSON save files.