Puzzle saving now checks for pre-existing files.
If the player tries to overwrite a pre-existing file, the game asks if they want to overwrite it. I need to add a canEdit check so that players cannot overwrite the original puzzles that came with the game. There is also a button to cancel trying to save if they want to do some more editing. Up next is actually saving the in-progress puzzle solution.