First version of cube saving possible.
The cubes that are built are now actually having their indices be saved. Up next is finding the minimum and maximum dimensions for each axis, and then reading in those values with the saved units to build a puzzle from scratch. Also need to make sure that the right puzzleUnits are deleted when a corresponding cube is deleted in build mode.