Loading json puzzles works!
Tried out a relatively fast implementation of loading from the Json files that are created in build mode. This is the start of making and saving puzzles faster in the game itself. Up next is adding a tracker to see if the remaining cubes are equal to the number of solution cubes to reach puzzle completion status.