Puzzle completion almost working
The game now detects when the puzzle being worked on has been completed. The player is able to retry the same puzzle or select a new one, but right now there is an issue with the silder's references being reset/lost/messed up when retrying a puzzle. Fixing that is the next todo item.