Puzzle replay errors handled
Something was up with trying to access deleted cubes and referencing the buttonHolder after resetting a puzzle upon completion, so creation of new puzzles and checking for a puzzle selector was cleaned up a bit. Also made the puzzle selection UI stop showing up after resetting a puzzle. Next up is making a clue/face editor that will allow me to manually choose faces for a puzzle to not be shown in a column. If every single column is always shown, the puzzles become too easy/formulaic, so I need a way to curate puzzles so that their entire solutions aren't always shown on all 3 axes.