Beginnings of clue editing functionality
Started the ClueEditorScript that will be used to hide/show specific clues in order to make puzzles more interesting. If all the faces of eavh cube is numbered, the puzzle is too easy and formulaic, so I need a way to manually select which faces are blank and which have numbers.