Pop SFX added to cube deletion. Start of clue hiding editor.
The cubes now make a popping sound when they are deleted. I've started a new script that will be used to possibly save a new version of a PuzzleSolution that will show whether a cube has its clues hidden or shown when it comes time to solve a puzzle. This is necessary to keep solving from being too easy/formulaic.