Single clue toggling working in clueEdit mode
The pair of faces on each cube can now be toggled on and off in clueEdit mode. Still need to find a way to save the changes to clues being hidden/shown so that they cna be set when a puzzle is created. Also need a way to set entire rows instead of just individual cubes to blank/numbered in puzzleMode.