Clue editor is working holy macaroni
Puzzles can now be made, clues can be edited, puzzles are essentially able to be made and saved in the form they'll functionally be in the final version of the puzzle. The clue ediotr is accessed when a puzzle is loaded in normally. The player can press E and click on faces in order to hide the clues, then press S and save a special version of the json file that has clueHiding info attached to it. Next big thing is fixing the slider issues at the start of their movement.