Added exit button and made several puzzles for testing.
The player can now exit the game by clicking the exit button in the menu screen. 7 puzzles and their clue-hidden versions were made in order to test puzzle solving. There is an oversight with flagging clues in that their values are revealed when flagged, even if the face was hidden, so that needs to be adressed asap.