New 1 pixel grid being used for cube textures
Set a new texture with 1 pixel cube edges, which ends up being 2 pixels wide when two cubes are next to each other, border around cubes. Still trying to get the updating of the center of the puzzle bounds to update when a cube is deleted. This might not even be necessary, a tutorial is more important.